+++ myems-api/excelexporters/virtualmetersaving.py
@@ -1,28 +1,21 @@
 import base64
-import uuid
 import os
-from decimal import Decimal
-from openpyxl.chart import PieChart, LineChart, BarChart, Reference
-from openpyxl.styles import PatternFill, Border, Side, Alignment, Font
-from openpyxl.drawing.image import Image
+import uuid
 from openpyxl import Workbook
+from openpyxl.chart import LineChart, Reference, Series
 from openpyxl.chart.label import DataLabelList
-import openpyxl.utils.cell as format_cell
+from openpyxl.drawing.image import Image
+from openpyxl.styles import PatternFill, Border, Side, Alignment, Font
 
 
 ########################################################################################################################
 # PROCEDURES
 # Step 1: Validate the report data
-# Step 2: Generate excel file
-# Step 3: Encode the excel file bytes to Base64
+# Step 2: Generate excel file from the report data
+# Step 3: Encode the excel file to Base64
 ########################################################################################################################
 
-
-def export(report,
-           name,
-           reporting_start_datetime_local,
-           reporting_end_datetime_local,
-           period_type):
+def export(report, name, reporting_start_datetime_local, reporting_end_datetime_local, period_type):
     ####################################################################################################################
     # Step 1: Validate the report data
     ####################################################################################################################
@@ -59,16 +52,11 @@ def export(report,
     return base64_message
 
 
-def generate_excel(report,
-                   name,
-                   reporting_start_datetime_local,
-                   reporting_end_datetime_local,
-                   period_type):
-    print(report)
+def generate_excel(report, name, reporting_start_datetime_local, reporting_end_datetime_local, period_type):
     wb = Workbook()
+    # todo
     ws = wb.active
     ws.title = "VirtualMeterSaving"
-
     # Row height
     ws.row_dimensions[1].height = 102
     for i in range(2, 2000 + 1):
@@ -145,646 +133,194 @@ def generate_excel(report,
     ws['E4'] = reporting_end_datetime_local
 
     if "reporting_period" not in report.keys() or \
-            "names" not in report['reporting_period'].keys() or len(report['reporting_period']['names']) == 0:
+            "values_saving" not in report['reporting_period'].keys() or \
+            len(report['reporting_period']['values_saving']) == 0:
         filename = str(uuid.uuid4()) + '.xlsx'
         wb.save(filename)
 
         return filename
 
     ####################################################################################################################
+    if "values_saving" not in report['reporting_period'].keys() or \
+            len(report['reporting_period']['values_saving']) == 0:
+        for i in range(6, 9 + 1):
+            ws.row_dimensions[i].height = 0.1
+    else:
+        ws['B6'].font = title_font
+        ws['B6'] = name + 'Reporting Period Saving'
 
-    current_row_number = 6
+        reporting_period_data = report['reporting_period']
+        category = report['virtual_meter']['energy_category_name']
+        ca_len = len(category)
 
+        ws.row_dimensions[7].height = 60
+        ws['B7'].fill = table_fill
+        ws['B7'].border = f_border
+
+        ws['B8'].font = title_font
+        ws['B8'].alignment = c_c_alignment
+        ws['B8'] = 'Saving'
+        ws['B8'].border = f_border
+
+        ws['B9'].font = title_font
+        ws['B9'].alignment = c_c_alignment
+        ws['B9'] = 'Increment Rate'
+        ws['B9'].border = f_border
+
+        col = 'B'
+
+        for i in range(0, ca_len):
+            col = chr(ord('C') + i)
+
+            ws[col + '7'].fill = table_fill
+            ws[col + '7'].font = name_font
+            ws[col + '7'].alignment = c_c_alignment
+            ws[col + '7'] = report['virtual_meter']['energy_category_name'] + " (" \
+                + report['virtual_meter']['unit_of_measure'] + ")"
+            ws[col + '7'].border = f_border
+
+            ws[col + '8'].font = name_font
+            ws[col + '8'].alignment = c_c_alignment
+            ws[col + '8'] = round(reporting_period_data['total_in_category_saving'], 2)
+            ws[col + '8'].border = f_border
+
+            ws[col + '9'].font = name_font
+            ws[col + '9'].alignment = c_c_alignment
+            ws[col + '9'] = str(round(reporting_period_data['increment_rate_saving'] * 100, 2)) + "%" \
+                if reporting_period_data['increment_rate_saving'] is not None else "-"
+            ws[col + '9'].border = f_border
+
+        # TCE TCO2E
+        end_col = col
+        # TCE
+        tce_col = chr(ord(end_col) + 1)
+        ws[tce_col + '7'].fill = table_fill
+        ws[tce_col + '7'].font = name_font
+        ws[tce_col + '7'].alignment = c_c_alignment
+        ws[tce_col + '7'] = 'Ton of Standard Coal (TCE) Saving'
+        ws[tce_col + '7'].border = f_border
+
+        ws[tce_col + '8'].font = name_font
+        ws[tce_col + '8'].alignment = c_c_alignment
+        ws[tce_col + '8'] = round(reporting_period_data['total_in_kgce_saving'] / 1000, 2)
+        ws[tce_col + '8'].border = f_border
+
+        ws[tce_col + '9'].font = name_font
+        ws[tce_col + '9'].alignment = c_c_alignment
+        ws[tce_col + '9'] = str(round(reporting_period_data['increment_rate_saving'] * 100, 2)) + "%" \
+            if reporting_period_data['increment_rate_saving'] is not None else "-"
+        ws[tce_col + '9'].border = f_border
+
+        # TCO2E
+        tco2e_col = chr(ord(end_col) + 2)
+        ws[tco2e_col + '7'].fill = table_fill
+        ws[tco2e_col + '7'].font = name_font
+        ws[tco2e_col + '7'].alignment = c_c_alignment
+        ws[tco2e_col + '7'] = 'Ton of Carbon Dioxide Emissions (TCO2E) Decreased'
+        ws[tco2e_col + '7'].border = f_border
+
+        ws[tco2e_col + '8'].font = name_font
+        ws[tco2e_col + '8'].alignment = c_c_alignment
+        ws[tco2e_col + '8'] = round(reporting_period_data['total_in_kgco2e_saving'] / 1000, 2)
+        ws[tco2e_col + '8'].border = f_border
+
+        ws[tco2e_col + '9'].font = name_font
+        ws[tco2e_col + '9'].alignment = c_c_alignment
+        ws[tco2e_col + '9'] = str(round(reporting_period_data['increment_rate_saving'] * 100, 2)) + "%" \
+            if reporting_period_data['increment_rate_saving'] is not None else "-"
+        ws[tco2e_col + '9'].border = f_border
+
+    ####################################################################################################################
     reporting_period_data = report['reporting_period']
-
-    has_names_data_flag = True
-
-    if "names" not in reporting_period_data.keys() or \
-            reporting_period_data['names'] is None or \
-            len(reporting_period_data['names']) == 0:
-        has_names_data_flag = False
-
-    if has_names_data_flag:
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)] = name + ' ' + 'Reporting Period Savings'
-
-        current_row_number += 1
-
-        ws.row_dimensions[current_row_number].height = 75
-        ws['B' + str(current_row_number)].fill = table_fill
-        ws['B' + str(current_row_number)].border = f_border
-
-        col = 'C'
-
-        ws[col + str(current_row_number)].fill = table_fill
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = \
-            reporting_period_data['names'] + "  (Baseline - Actual) (" + reporting_period_data['units'] + ")"
-
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].fill = table_fill
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = 'Ton of Standard Coal  (Baseline - Actual) (TCE)'
-
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].fill = table_fill
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = 'Ton of Carbon Dioxide Emissions  (Baseline - Actual) (TCO2E)'
-
-        current_row_number += 1
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)].alignment = c_c_alignment
-        ws['B' + str(current_row_number)].border = f_border
-        ws['B' + str(current_row_number)] = 'Savings'
-
-        col = 'C'
-
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = round(reporting_period_data['subtotals_saving'], 2)
-
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = round(reporting_period_data['total_in_kgce_saving'] / 1000, 2)
-
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = round(reporting_period_data['total_in_kgco2e_saving'] / 1000, 2)
-
-        current_row_number += 1
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)].alignment = c_c_alignment
-        ws['B' + str(current_row_number)].border = f_border
-        ws['B' + str(current_row_number)] = 'Increment Rate'
-
-        col = 'C'
-
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = str(
-            round(reporting_period_data['increment_rates_saving'] * 100, 2)) + '%' \
-            if reporting_period_data['increment_rates_saving'] is not None else '-'
-
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = str(
-            round(reporting_period_data['increment_rate_in_kgce_saving'] * 100, 2)) + '%' \
-            if reporting_period_data['increment_rate_in_kgce_saving'] is not None else '-'
-
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].font = name_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = str(
-            round(reporting_period_data['increment_rate_in_kgco2e_saving'] * 100, 2)) + '%' \
-            if reporting_period_data['increment_rate_in_kgco2e_saving'] is not None else '-'
-
-        current_row_number += 2
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)] = name + ' ' + 'Ton of Standard Coal(TCE) by Energy Category'
-
-        current_row_number += 1
-        table_start_row_number = current_row_number
-        chart_start_row_number = current_row_number
-
-        ws.row_dimensions[current_row_number].height = 60
-        ws['B' + str(current_row_number)].fill = table_fill
-        ws['B' + str(current_row_number)].border = f_border
-
-        ws['C' + str(current_row_number)].fill = table_fill
-        ws['C' + str(current_row_number)].font = name_font
-        ws['C' + str(current_row_number)].alignment = c_c_alignment
-        ws['C' + str(current_row_number)].border = f_border
-        ws['C' + str(current_row_number)] = 'Savings'
-
-        ws['D' + str(current_row_number)].fill = table_fill
-        ws['D' + str(current_row_number)].font = name_font
-        ws['D' + str(current_row_number)].alignment = c_c_alignment
-        ws['D' + str(current_row_number)].border = f_border
-        ws['D' + str(current_row_number)] = 'Ton of Standard Coal(TCE) by Energy Category'
-
-        current_row_number += 1
-
-        subtotals_in_kgce_saving_sum = reporting_period_data['subtotals_in_kgce_saving']
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)].alignment = c_c_alignment
-        ws['B' + str(current_row_number)].border = f_border
-        ws['B' + str(current_row_number)] = reporting_period_data['names']
-
-        ws['C' + str(current_row_number)].font = name_font
-        ws['C' + str(current_row_number)].alignment = c_c_alignment
-        ws['C' + str(current_row_number)].border = f_border
-        ws['C' + str(current_row_number)] = round(reporting_period_data['subtotals_in_kgce_saving'] / 1000, 3)
-
-        ws['D' + str(current_row_number)].font = name_font
-        ws['D' + str(current_row_number)].alignment = c_c_alignment
-        ws['D' + str(current_row_number)].border = f_border
-        ws['D' + str(current_row_number)] = str(round(reporting_period_data['subtotals_in_kgce_saving'] /
-                                                      subtotals_in_kgce_saving_sum * 100, 2)) + '%' \
-            if abs(subtotals_in_kgce_saving_sum) > 0 else '-'
-
-        current_row_number += 5
-
-        table_end_row_number = current_row_number - 1
-
-        pie = PieChart()
-        pie.title = name + ' ' + 'Ton of Standard Coal(TCE) by Energy Category'
-        labels = Reference(ws, min_col=2, min_row=table_start_row_number + 1, max_row=table_end_row_number)
-        pie_data = Reference(ws, min_col=3, min_row=table_start_row_number, max_row=table_end_row_number)
-        pie.add_data(pie_data, titles_from_data=True)
-        pie.set_categories(labels)
-        pie.height = 7.25
-        pie.width = 9
-        s1 = pie.series[0]
-        s1.dLbls = DataLabelList()
-        s1.dLbls.showCatName = False
-        s1.dLbls.showVal = True
-        s1.dLbls.showPercent = True
-        ws.add_chart(pie, 'E' + str(chart_start_row_number))
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)] = name + ' ' + 'Ton of Carbon Dioxide Emissions(TCO2E) by Energy Category'
-
-        current_row_number += 1
-        table_start_row_number = current_row_number
-        chart_start_row_number = current_row_number
-
-        ws.row_dimensions[current_row_number].height = 60
-        ws['B' + str(current_row_number)].fill = table_fill
-        ws['B' + str(current_row_number)].border = f_border
-
-        ws['C' + str(current_row_number)].fill = table_fill
-        ws['C' + str(current_row_number)].font = name_font
-        ws['C' + str(current_row_number)].alignment = c_c_alignment
-        ws['C' + str(current_row_number)].border = f_border
-        ws['C' + str(current_row_number)] = 'Savings'
-
-        ws['D' + str(current_row_number)].fill = table_fill
-        ws['D' + str(current_row_number)].font = name_font
-        ws['D' + str(current_row_number)].alignment = c_c_alignment
-        ws['D' + str(current_row_number)].border = f_border
-        ws['D' + str(current_row_number)] = 'Ton of Carbon Dioxide Emissions(TCO2E) by Energy Category'
-
-        current_row_number += 1
-
-        subtotals_in_kgco2e_saving_sum = reporting_period_data['subtotals_in_kgco2e_saving']
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)].alignment = c_c_alignment
-        ws['B' + str(current_row_number)].border = f_border
-        ws['B' + str(current_row_number)] = reporting_period_data['names']
-
-        ws['C' + str(current_row_number)].font = name_font
-        ws['C' + str(current_row_number)].alignment = c_c_alignment
-        ws['C' + str(current_row_number)].border = f_border
-        ws['C' + str(current_row_number)] = round(reporting_period_data['subtotals_in_kgco2e_saving'] / 1000, 3)
-
-        ws['D' + str(current_row_number)].font = name_font
-        ws['D' + str(current_row_number)].alignment = c_c_alignment
-        ws['D' + str(current_row_number)].border = f_border
-        ws['D' + str(current_row_number)] = str(round(reporting_period_data['subtotals_in_kgco2e_saving'] /
-                                                      subtotals_in_kgco2e_saving_sum * 100, 2)) + '%' \
-            if abs(subtotals_in_kgco2e_saving_sum) > 0 else '-'
-
-        current_row_number += 1
-
-        table_end_row_number = current_row_number - 1
-
-        current_row_number += 4
-
-        pie = PieChart()
-        pie.title = name + ' ' + 'Ton of Carbon Dioxide Emissions(TCO2E) by Energy Category'
-        labels = Reference(ws, min_col=2, min_row=table_start_row_number + 1, max_row=table_end_row_number)
-        pie_data = Reference(ws, min_col=3, min_row=table_start_row_number, max_row=table_end_row_number)
-        pie.add_data(pie_data, titles_from_data=True)
-        pie.set_categories(labels)
-        pie.height = 7.25
-        pie.width = 9
-        s1 = pie.series[0]
-        s1.dLbls = DataLabelList()
-        s1.dLbls.showCatName = False
-        s1.dLbls.showVal = True
-        s1.dLbls.showPercent = True
-        ws.add_chart(pie, 'E' + str(chart_start_row_number))
-
-    ####################################################################################################################
-
-    has_values_saving_data = True
-    has_timestamps_data = True
-
-    if 'values_saving' not in reporting_period_data.keys() or \
-            reporting_period_data['values_saving'] is None or \
-            len(reporting_period_data['values_saving']) == 0:
-        has_values_saving_data = False
-
-    if 'timestamps' not in reporting_period_data.keys() or \
-            reporting_period_data['timestamps'] is None or \
-            len(reporting_period_data['timestamps']) == 0 or \
-            len(reporting_period_data['timestamps'][0]) == 0:
-        has_timestamps_data = False
-
-    current_row_number += 1
-    chart_start_row_number = current_row_number
-    if has_values_saving_data and has_timestamps_data:
-        time = reporting_period_data['timestamps']
-        parameters_names_len = len(report['parameters']['names'])
-        parameters_parameters_datas_len = 0
-
-        ws['B' + str(current_row_number - 1)].font = title_font
-        ws['B' + str(current_row_number - 1)] = name + ' ' + 'Detailed Data'
-
-        for i in range(0, parameters_names_len):
-            if len(report['parameters']['timestamps'][i]) == 0:
-                continue
-            parameters_parameters_datas_len += 1
-
-        current_row_number = current_row_number + (1 + parameters_parameters_datas_len) * 6 + 2
-        table_start_row_number = current_row_number
-
-        ws.row_dimensions[current_row_number].height = 60
-        ws['B' + str(current_row_number)].fill = table_fill
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)].alignment = c_c_alignment
-        ws['B' + str(current_row_number)].border = f_border
-        ws['B' + str(current_row_number)] = 'Datetime'
-
-        col = 'C'
-
-        ws[col + str(current_row_number)].fill = table_fill
-        ws[col + str(current_row_number)].font = title_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = \
-            reporting_period_data['names'] + ':Actual' + " (" + reporting_period_data['units'] + ")"
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].fill = table_fill
-        ws[col + str(current_row_number)].font = title_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = \
-            reporting_period_data['names'] + ':Baseline' + " (" + reporting_period_data['units'] + ")"
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].fill = table_fill
-        ws[col + str(current_row_number)].font = title_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = \
-            reporting_period_data['names'] + ':Saving' + " (" + reporting_period_data['units'] + ")"
-
-        current_row_number += 1
-        print(time)
-        for i in range(0, len(time)):
-            ws['B' + str(current_row_number)].font = title_font
-            ws['B' + str(current_row_number)].alignment = c_c_alignment
-            ws['B' + str(current_row_number)].border = f_border
-            ws['B' + str(current_row_number)] = time[i]
-
-            col = 'C'
-
-            ws[col + str(current_row_number)].font = title_font
-            ws[col + str(current_row_number)].alignment = c_c_alignment
-            ws[col + str(current_row_number)].border = f_border
-            ws[col + str(current_row_number)] = round(reporting_period_data['values_actual'][i], 2) \
-                if reporting_period_data['values_actual'] is not None else 0.00
-            col = chr(ord(col) + 1)
-
-            ws[col + str(current_row_number)].font = title_font
-            ws[col + str(current_row_number)].alignment = c_c_alignment
-            ws[col + str(current_row_number)].border = f_border
-            ws[col + str(current_row_number)] = round(reporting_period_data['values_baseline'][i], 2) \
-                if reporting_period_data['values_baseline'][i] is not None else 0.00
-            col = chr(ord(col) + 1)
-
-            ws[col + str(current_row_number)].font = title_font
-            ws[col + str(current_row_number)].alignment = c_c_alignment
-            ws[col + str(current_row_number)].border = f_border
-            ws[col + str(current_row_number)] = round(reporting_period_data['values_saving'][i], 2) \
-                if reporting_period_data['values_saving'][i] is not None else 0.00
-
-            current_row_number += 1
-
-        table_end_row_number = current_row_number - 1
-
-        ws['B' + str(current_row_number)].font = title_font
-        ws['B' + str(current_row_number)].alignment = c_c_alignment
-        ws['B' + str(current_row_number)].border = f_border
-        ws['B' + str(current_row_number)] = 'Subtotal'
-
-        col = 'C'
-
-        ws[col + str(current_row_number)].font = title_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = round(reporting_period_data['subtotals_actual'], 2)
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].font = title_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = round(reporting_period_data['subtotals_baseline'], 2)
-        col = chr(ord(col) + 1)
-
-        ws[col + str(current_row_number)].font = title_font
-        ws[col + str(current_row_number)].alignment = c_c_alignment
-        ws[col + str(current_row_number)].border = f_border
-        ws[col + str(current_row_number)] = round(reporting_period_data['subtotals_saving'], 2)
-        current_row_number += 2
-
-        format_time_width_number = 1.0
-        min_len_number = 1.0
-        min_width_number = 11.0
-
-        if period_type == 'hourly':
-            format_time_width_number = 4.0
-            min_len_number = 2
-            min_width_number = 12.0
-        elif period_type == 'daily':
-            format_time_width_number = 2.5
-            min_len_number = 4
-            min_width_number = 14.0
-        elif period_type == 'monthly':
-            format_time_width_number = 2.1
-            min_len_number = 4
-            min_width_number = 12.4
-        elif period_type == 'yearly':
-            format_time_width_number = 1.5
-            min_len_number = 5
-            min_width_number = 11.5
-
-        line = LineChart()
-        line.title = 'Reporting Period Savings - ' + \
-                     reporting_period_data['names'] + " (" + reporting_period_data['units'] + ")"
-        labels = Reference(ws, min_col=2, min_row=table_start_row_number + 1, max_row=table_end_row_number)
-        line_data = Reference(ws, min_col=3, max_col=4, min_row=table_start_row_number, max_row=table_end_row_number)
-        line.add_data(line_data, titles_from_data=True)
-        line.set_categories(labels)
-        line_data = line.series[0]
-        line_data.marker.symbol = "circle"
-        line_data.smooth = True
-        line.height = 8.25
-        line.width = format_time_width_number * len(time) if len(time) > min_len_number else min_width_number
-        if line.width > 24:
-            line.width = 24
-        line.x_axis.crosses = 'min'
-        line.dLbls = DataLabelList()
-        line.dLbls.dLblPos = 't'
-        line.dLbls.showVal = True
-        line.dLbls.showPercent = False
-        chart_col = 'B'
-        chart_cell = chart_col + str(chart_start_row_number)
-        ws.add_chart(line, chart_cell)
-
-        bar = BarChart()
-        bar.type = "col"
-        bar.style = 10
-        bar.height = 8.25
-        bar.width = line.width
-        bar.title = 'Reporting Period Savings - ' + \
-                    reporting_period_data['names'] + " (" + reporting_period_data['units'] + ")"
-        labels = Reference(ws, min_col=2, min_row=table_start_row_number + 1, max_row=table_end_row_number)
-        bar_data = Reference(ws, min_col=5, min_row=table_start_row_number, max_row=table_end_row_number)
-        bar.add_data(bar_data, titles_from_data=True)
-        bar.set_categories(labels)
-        bar.dLbls = DataLabelList()
-        bar.dLbls.showVal = True
-        bar.dLbls.showPercent = False
-        chart_col = 'B'
-        chart_cell = chart_col + str(chart_start_row_number + 1 * 6 + 1)
-        bar.shape = 4
-        ws.add_chart(bar, chart_cell)
-
-    ####################################################################################################################
-    has_parameters_names_and_timestamps_and_values_data = True
-
-    current_sheet_parameters_row_number = chart_start_row_number + 1 * 6
-    if 'parameters' not in report.keys() or \
-            report['parameters'] is None or \
-            'names' not in report['parameters'].keys() or \
-            report['parameters']['names'] is None or \
-            len(report['parameters']['names']) == 0 or \
-            'timestamps' not in report['parameters'].keys() or \
-            report['parameters']['timestamps'] is None or \
-            len(report['parameters']['timestamps']) == 0 or \
-            'values' not in report['parameters'].keys() or \
-            report['parameters']['values'] is None or \
-            len(report['parameters']['values']) == 0 or \
-            timestamps_data_all_equal_0(report['parameters']['timestamps']):
-        has_parameters_names_and_timestamps_and_values_data = False
-    if has_parameters_names_and_timestamps_and_values_data:
-
-        ################################################################################################################
-        # new worksheet
-        ################################################################################################################
-
-        parameters_data = report['parameters']
-
-        parameters_names_len = len(parameters_data['names'])
-
-        file_name = ws.title
-        parameters_ws = wb.create_sheet(file_name + 'Parameters')
-
-        parameters_timestamps_data_max_len = \
-            get_parameters_timestamps_lists_max_len(list(parameters_data['timestamps']))
-
-        # Row height
-        parameters_ws.row_dimensions[1].height = 102
-        for i in range(2, 7 + 1):
-            parameters_ws.row_dimensions[i].height = 42
-
-        for i in range(8, parameters_timestamps_data_max_len + 10):
-            parameters_ws.row_dimensions[i].height = 60
-
-        # Col width
-        parameters_ws.column_dimensions['A'].width = 1.5
-
-        parameters_ws.column_dimensions['B'].width = 25.0
-
-        for i in range(3, 12 + parameters_names_len * 3):
-            parameters_ws.column_dimensions[format_cell.get_column_letter(i)].width = 15.0
-
-        # Img
-        img = Image("excelexporters/myems.png")
-        parameters_ws.add_image(img, 'A1')
-
-        # Title
-        parameters_ws['B3'].alignment = b_r_alignment
-        parameters_ws['B3'] = 'Name:'
-        parameters_ws['C3'].border = b_border
-        parameters_ws['C3'].alignment = b_c_alignment
-        parameters_ws['C3'] = name
-
-        parameters_ws['D3'].alignment = b_r_alignment
-        parameters_ws['D3'] = 'Period:'
-        parameters_ws['E3'].border = b_border
-        parameters_ws['E3'].alignment = b_c_alignment
-        parameters_ws['E3'] = period_type
-
-        parameters_ws['B4'].alignment = b_r_alignment
-        parameters_ws['B4'] = 'Reporting Start Datetime:'
-        parameters_ws['C4'].border = b_border
-        parameters_ws['C4'].alignment = b_c_alignment
-        parameters_ws['C4'] = reporting_start_datetime_local
-
-        parameters_ws['D4'].alignment = b_r_alignment
-        parameters_ws['D4'] = 'Reporting End Datetime:'
-        parameters_ws['E4'].border = b_border
-        parameters_ws['E4'].alignment = b_c_alignment
-        parameters_ws['E4'] = reporting_end_datetime_local
-
-        parameters_ws_current_row_number = 6
-
-        parameters_ws['B' + str(parameters_ws_current_row_number)].font = title_font
-        parameters_ws['B' + str(parameters_ws_current_row_number)] = name + ' ' + 'Parameters'
-
-        parameters_ws_current_row_number += 1
-
-        parameters_table_start_row_number = parameters_ws_current_row_number
-
-        parameters_ws.row_dimensions[parameters_ws_current_row_number].height = 80
-
-        parameters_ws_current_row_number += 1
-
-        table_current_col_number = 2
-
-        for i in range(0, parameters_names_len):
-
-            if len(parameters_data['timestamps'][i]) == 0:
-                continue
-
-            col = format_cell.get_column_letter(table_current_col_number)
-
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)].fill = table_fill
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)].border = f_border
-
-            col = format_cell.get_column_letter(table_current_col_number + 1)
-
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)].fill = table_fill
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)].border = f_border
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)].font = name_font
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)].alignment = c_c_alignment
-            parameters_ws[col + str(parameters_ws_current_row_number - 1)] = parameters_data['names'][i]
-
-            table_current_row_number = parameters_ws_current_row_number
-
-            for j, value in enumerate(list(parameters_data['timestamps'][i])):
-                col = format_cell.get_column_letter(table_current_col_number)
-
-                parameters_ws[col + str(table_current_row_number)].border = f_border
-                parameters_ws[col + str(table_current_row_number)].font = title_font
-                parameters_ws[col + str(table_current_row_number)].alignment = c_c_alignment
-                parameters_ws[col + str(table_current_row_number)] = value
-
-                col = format_cell.get_column_letter(table_current_col_number + 1)
-
-                parameters_ws[col + str(table_current_row_number)].border = f_border
-                parameters_ws[col + str(table_current_row_number)].font = title_font
-                parameters_ws[col + str(table_current_row_number)].alignment = c_c_alignment
-                parameters_ws[col + str(table_current_row_number)] = round(parameters_data['values'][i][j], 2)
-
-                table_current_row_number += 1
-
-            table_current_col_number = table_current_col_number + 3
-
-        ################################################################################################################
-        # parameters chart and parameters table
-        ################################################################################################################
-
-        ws['B' + str(current_sheet_parameters_row_number)].font = title_font
-        ws['B' + str(current_sheet_parameters_row_number)] = name + ' ' + 'Parameters'
-
-        current_sheet_parameters_row_number += 1
-
-        chart_start_row_number = current_sheet_parameters_row_number
-
-        col_index = 0
-
-        for i in range(0, parameters_names_len):
-
-            if len(parameters_data['timestamps'][i]) == 0:
-                continue
+    category = report['virtual_meter']['energy_category_name']
+    ca_len = len(category)
+    times = reporting_period_data['timestamps']
+
+    if "values_saving" not in reporting_period_data.keys() or len(reporting_period_data['values_saving']) == 0:
+        for i in range(11, 43 + 1):
+            ws.row_dimensions[i].height = 0.0
+    else:
+        ws['B11'].font = title_font
+        ws['B11'] = name + 'Detailed Data'
+
+        ws.row_dimensions[18].height = 60
+        ws['B18'].fill = table_fill
+        ws['B18'].font = title_font
+        ws['B18'].border = f_border
+        ws['B18'].alignment = c_c_alignment
+        ws['B18'] = 'Datetime'
+        time = times
+        has_data = False
+        max_row = 0
+        if len(time) > 0:
+            has_data = True
+            max_row = 18 + len(time)
+
+        if has_data:
+
+            end_data_row_number = 19
+
+            for i in range(0, len(time)):
+                col = 'B'
+                end_data_row_number = 19 + i
+                row = str(end_data_row_number)
+
+                ws[col + row].font = title_font
+                ws[col + row].alignment = c_c_alignment
+                ws[col + row] = time[i]
+                ws[col + row].border = f_border
+
+            ws['B' + str(end_data_row_number + 1)].font = title_font
+            ws['B' + str(end_data_row_number + 1)].alignment = c_c_alignment
+            ws['B' + str(end_data_row_number + 1)] = 'Total'
+            ws['B' + str(end_data_row_number + 1)].border = f_border
+
+            for i in range(0, ca_len):
+
+                col = chr(ord('C') + i)
+
+                ws[col + '18'].fill = table_fill
+                ws[col + '18'].font = title_font
+                ws[col + '18'].alignment = c_c_alignment
+                ws[col + '18'] = report['virtual_meter']['energy_category_name'] + " (" \
+                    + report['virtual_meter']['unit_of_measure'] + ")"
+                ws[col + '18'].border = f_border
+
+                time = times
+                time_len = len(time)
+
+                for j in range(0, time_len):
+                    row = str(19 + j)
+
+                    ws[col + row].font = title_font
+                    ws[col + row].alignment = c_c_alignment
+                    ws[col + row] = round(reporting_period_data['values_saving'][j], 2)
+                    ws[col + row].border = f_border
+
+                ws[col + str(end_data_row_number + 1)].font = title_font
+                ws[col + str(end_data_row_number + 1)].alignment = c_c_alignment
+                ws[col + str(end_data_row_number + 1)] = round(reporting_period_data['total_in_category_saving'], 2)
+                ws[col + str(end_data_row_number + 1)].border = f_border
 
             line = LineChart()
-            data_col = 3 + col_index * 3
-            labels_col = 2 + col_index * 3
-            col_index += 1
-            line.title = 'Parameters - ' + \
-                         parameters_ws.cell(row=parameters_table_start_row_number, column=data_col).value
-            labels = Reference(parameters_ws, min_col=labels_col, min_row=parameters_table_start_row_number + 1,
-                               max_row=(len(parameters_data['timestamps'][i]) + parameters_table_start_row_number))
-            line_data = Reference(parameters_ws, min_col=data_col, min_row=parameters_table_start_row_number,
-                                  max_row=(len(parameters_data['timestamps'][i]) + parameters_table_start_row_number))
-            line.add_data(line_data, titles_from_data=True)
+            labels = Reference(ws, min_col=2, min_row=19, max_row=max_row)
+            line_data = Reference(ws, min_col=3, min_row=18, max_row=max_row)
+            line.series.append(Series(line_data, title_from_data=True))
             line.set_categories(labels)
             line_data = line.series[0]
             line_data.marker.symbol = "circle"
             line_data.smooth = True
             line.x_axis.crosses = 'min'
-            line.height = 8.25
-            line.width = 24
+            line.title = 'Reporting Period Saving - ' + report['virtual_meter']['energy_category_name'] + \
+                         " (" + report['virtual_meter']['unit_of_measure'] + ")"
             line.dLbls = DataLabelList()
             line.dLbls.dLblPos = 't'
-            line.dLbls.showVal = False
-            line.dLbls.showPercent = False
-            chart_col = 'B'
-            chart_cell = chart_col + str(chart_start_row_number)
-            chart_start_row_number += 6
-            ws.add_chart(line, chart_cell)
-
-        current_sheet_parameters_row_number = chart_start_row_number
-
-        current_sheet_parameters_row_number += 1
+            line.dLbls.showVal = True
+            line.height = 8.25
+            line.width = 24
+            ws.add_chart(line, "B12")
 
     filename = str(uuid.uuid4()) + '.xlsx'
     wb.save(filename)
 
     return filename
-
-
-def sum_list(lists):
-    total = Decimal(0)
-    for i in range(0, len(lists)):
-        total += lists[i]
-    return total
-
-
-def get_parameters_timestamps_lists_max_len(parameters_timestamps_lists):
-    max_len = 0
-    for i, value in enumerate(list(parameters_timestamps_lists)):
-        if len(value) > max_len:
-            max_len = len(value)
-
-    return max_len
-
-
-def timestamps_data_all_equal_0(lists):
-    for i, value in enumerate(list(lists)):
-        if len(value) > 0:
-            return False
-
-    return True